chore(overview): cleanup validation code
