+++ shared/src/ProjectTreeTable/context/CellEditingContext.tsx
@@ -7,6 +7,7 @@ import useUpdateOverview, {
 import { useProjectTableContext } from './ProjectTableContext'
 import { AttributeData } from '../types'
 import { toast } from 'react-toastify'
+import useValidateUpdates from '../hooks/useValidateUpdates'
 
 export interface CellEditingContextType {
   editingCellId: CellId | null
@@ -25,78 +26,22 @@ export const CellEditingProvider: React.FC<{ children: ReactNode }> = ({ childre
   const isEditing = useCallback((id: CellId) => id === editingCellId, [editingCellId])
 
   const { updateEntities, inheritFromParent } = useUpdateOverview()
-  const { attribFields } = useProjectTableContext()
-  const validateUpdateEntities: UpdateTableEntities = async (entities = []) => {
+
+  const validateUpdateEntities = useValidateUpdates()
+
+  const handleUpdateEntities: UpdateTableEntities = async (entities = []) => {
     try {
-      // first validate the values are correct
-      for (const { isAttrib, value: rawValue, field } of entities) {
-        if (!isAttrib) continue
-        const attribute = attribFields.find((attr) => attr.name === field)
-        if (!attribute) continue
+      // validate the entities before updating
+      validateUpdateEntities(entities)
 
-        // coerce numeric strings into numbers for integer/float types or fail
-        let value: any = rawValue
-        const { type } = attribute.data
-        if (type === 'integer' || type === 'float') {
-          if (typeof rawValue === 'string') {
-            // empty or non‑numeric strings are invalid
-            if (rawValue.trim() === '' || isNaN(Number(rawValue))) {
-              throw new Error(`“${field}” must be a valid number`)
-            }
-            value = type === 'integer' ? parseInt(rawValue, 10) : parseFloat(rawValue)
-          } else if (typeof rawValue !== 'number') {
-            // any other type is invalid
-            throw new Error(`“${field}” must be a valid number`)
-          }
-        }
-
-        // collect numeric rules from attribute.data
-        const validationKeys: (keyof AttributeData)[] = [
-          'ge',
-          'gt',
-          'le',
-          'lt',
-          'minLength',
-          'maxLength',
-          'minItems',
-          'maxItems',
-        ]
-        const validationValues = (
-          Object.entries(attribute.data) as [keyof AttributeData, any][]
-        ).reduce((acc, [key, v]) => {
-          if (validationKeys.includes(key)) acc[key] = v as number
-          return acc
-        }, {} as Record<keyof AttributeData, number>)
-
-        const { ge, gt, le, lt, minLength, maxLength, minItems, maxItems } = validationValues
-        const pattern = attribute.data.regex
-
-        if (typeof value === 'number') {
-          if (ge != null && value < ge) throw new Error(`“${field}” must be ≥ ${ge}`)
-          if (gt != null && value <= gt) throw new Error(`“${field}” must be > ${gt}`)
-          if (le != null && value > le) throw new Error(`“${field}” must be ≤ ${le}`)
-          if (lt != null && value >= lt) throw new Error(`“${field}” must be < ${lt}`)
-        } else if (typeof value === 'string') {
-          if (minLength != null && value.length < minLength)
-            throw new Error(`“${field}” length must be ≥ ${minLength}`)
-          if (maxLength != null && value.length > maxLength)
-            throw new Error(`“${field}” length must be ≤ ${maxLength}`)
-          if (pattern && !new RegExp(pattern).test(value))
-            throw new Error(`“${field}” must match pattern ${pattern}`)
-        } else if (Array.isArray(value)) {
-          if (minItems != null && value.length < minItems)
-            throw new Error(`“${field}” items must be ≥ ${minItems}`)
-          if (maxItems != null && value.length > maxItems)
-            throw new Error(`“${field}” items must be ≤ ${maxItems}`)
-        }
-      }
+      // if validation passes, update the entities
+      return await updateEntities(entities)
     } catch (error: any) {
       // if validation fails, show a toast and return
       toast.error(error.message)
-      return Promise.resolve()
+
+      return Promise.reject(error)
     }
-    // all good – forward to the real updater
-    return updateEntities(entities)
   }
 
   const value = useMemo(
@@ -104,7 +49,7 @@ export const CellEditingProvider: React.FC<{ children: ReactNode }> = ({ childre
       editingCellId,
       setEditingCellId,
       isEditing,
-      updateEntities: validateUpdateEntities,
+      updateEntities: handleUpdateEntities,
       inheritFromParent,
     }),
     [editingCellId, isEditing, updateEntities],
